refactor: carry out listener function to module
