+++ src/js/modules/languageAndCapitalaizeSwitch.js
@@ -3,7 +3,7 @@ const languageAndCapitalaizeSwitch = (e, listOfKeys, currentLang, capsState) => 
     listOfKeys[29].classList.toggle('caps-active');
     listOfKeys.forEach(x => {
       if (currentLang === 'ru' && !x.classList.contains('caps-inactive-ru')) {
-        if (!capsState) {
+        if (capsState) {
           const item = x;
           item.innerHTML = x.dataset.ru_upper;
         } else {
@@ -11,7 +11,7 @@ const languageAndCapitalaizeSwitch = (e, listOfKeys, currentLang, capsState) => 
           item.innerHTML = x.dataset.ru_lower;
         }
       } else if (currentLang === 'en' && !x.classList.contains('caps-inactive-en')) {
-        if (!capsState) {
+        if (capsState) {
           const item = x;
           item.innerHTML = x.dataset.en_upper;
         } else {
@@ -29,10 +29,17 @@ const languageAndCapitalaizeSwitch = (e, listOfKeys, currentLang, capsState) => 
     if (currentLang === 'ru') {
       localStorage.setItem('lang', 'ru');
     }
+
     if (capsState && currentLang === 'en') {
       listOfKeys.forEach(x => {
         const item = x;
-        item.innerHTML = x.dataset.en_upper;
+        item.innerHTML = x.dataset.en_lower;
+      });
+      listOfKeys.forEach(x => {
+        const item = x;
+        if (!x.classList.contains('caps-inactive-en')) {
+          item.innerHTML = x.dataset.en_upper;
+        }
       });
     }
     if (!capsState && currentLang === 'en') {
@@ -44,7 +51,13 @@ const languageAndCapitalaizeSwitch = (e, listOfKeys, currentLang, capsState) => 
     if (capsState && currentLang === 'ru') {
       listOfKeys.forEach(x => {
         const item = x;
-        item.innerHTML = x.dataset.ru_upper;
+        item.innerHTML = x.dataset.ru_lower;
+      });
+      listOfKeys.forEach(x => {
+        const item = x;
+        if (!x.classList.contains('caps-inactive-ru')) {
+          item.innerHTML = x.dataset.ru_upper;
+        }
       });
     }
     if (!capsState && currentLang === 'ru') {
